Refactor: 캘린더 화면 UI 리팩토링
CalendarUtil 파일 내 주요 Composable들 별도 파일로 분리
화면에 사용되던 상수들 CalendarDesignToken object에 모아서 관리
