+++ feature/main/src/main/java/kr/co/main/calendar/common/CalendarDesignToken.kt
@@ -1,0 +1,28 @@
+package kr.co.main.calendar.common
+
+internal object CalendarDesignToken {
+    const val ROUNDED_CORNER_RADIUS = 12
+
+    // 캘린더 화면 common ui
+    const val FAB_ICON_SIZE = 20
+    const val SKY_ICON_SIZE = 20
+    const val CROP_COLOR_SHAPE_SIZE = 8
+
+    // 농작업 일정 캘린더
+    const val FARM_WORK_ITEM_HEIGHT = 20
+
+    // 메인 캘린더
+    const val SCHEDULE_ITEM_HEIGHT = 20
+    const val DIARY_ITEM_ICON_SIZE = 24
+
+    // 영농일지 추가 화면 이미지 입력
+    const val IMAGE_SIZE = 66
+    const val ADD_IMAGE_BUTTON_SIZE = 66
+    const val DELETE_IMAGE_BUTTON_SIZE = 16
+    const val DELETE_IMAGE_BUTTON_OFFSET = 2
+
+    // 영농일지 추가 화면 작업 입력
+    const val WORK_TYPE_SPINNER_ICON_SIZE = 20
+    const val WORK_TYPE_SPINNER_DROP_DOWN_OFFSET = 50
+    const val WORK_TYPE_SPINNER_DROP_DOWN_HEIGHT = 300
+}
